requirements.txt
Telethon~=1.37.0
aiogram~=3.14.0
python-dotenv~=1.0.1